# NGGIS-Connector-Source-20260918-085605/backup-v8-swagger-20260917-143322/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1",
    "title": "National Grid GIS Enhanced V7",
    "description": "Provides read-only enhanced access to structured National Grid GIS PROD data.",
    "x-ms-api-annotation": {
      "status": "Preview"
    },
    "x-ms-keywords": [
      "National Grid",
      "GIS",
      "ArcGIS"
    ],
    "contact": {
      "name": "National Grid",
      "email": "Grant.Lella@nationalgrid.com"
    }
  },
  "host": "127.0.0.1:5008",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/gis/geoprocessing/geometry/{operation}": {
      "post": {
        "operationId": "RunEnterpriseGeometryOperation",
        "summary": "Run ArcGIS Enterprise geometry operation",
        "description": "Run ArcGIS Enterprise geometry operation",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "operation",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Geometry operation",
            "enum": [
              "areasAndLengths",
              "autoComplete",
              "buffer",
              "convexHull",
              "cut",
              "densify",
              "difference",
              "distance",
              "findTransformations",
              "fromGeoCoordinateString",
              "generalize",
              "intersect",
              "labelPoints",
              "lengths",
              "offset",
              "project",
              "relation",
              "reshape",
              "simplify",
              "toGeoCoordinateString",
              "trimExtend",
              "union"
            ]
          },
          {
            "name": "parameters",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "ArcGIS REST parameters as JSON. Tokens are added internally."
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/execute": {
      "post": {
        "operationId": "ExecuteEnterpriseGpTask",
        "summary": "Execute synchronous Enterprise GP task",
        "description": "Execute synchronous Enterprise GP task",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP task URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer task URL."
          },
          {
            "name": "parameters",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "ArcGIS REST parameters as JSON. Tokens are added internally."
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/submitJob": {
      "post": {
        "operationId": "SubmitEnterpriseGpJob",
        "summary": "Submit asynchronous Enterprise GP job",
        "description": "Submit asynchronous Enterprise GP job",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP task URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer task URL."
          },
          {
            "name": "parameters",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "ArcGIS REST parameters as JSON. Tokens are added internally."
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/job": {
      "get": {
        "operationId": "GetEnterpriseGpJob",
        "summary": "Get Enterprise GP job status",
        "description": "Get Enterprise GP job status",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP task URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer task URL."
          },
          {
            "name": "jobId",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/result": {
      "get": {
        "operationId": "GetEnterpriseGpResult",
        "summary": "Get Enterprise GP job result",
        "description": "Get Enterprise GP job result",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP task URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer task URL."
          },
          {
            "name": "jobId",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Job ID"
          },
          {
            "name": "resultName",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Result name"
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/cancel": {
      "post": {
        "operationId": "CancelEnterpriseGpJob",
        "summary": "Cancel Enterprise GP job",
        "description": "Cancel Enterprise GP job",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP task URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer task URL."
          },
          {
            "name": "jobId",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    },
    "/gis/geoprocessing/gp/resource": {
      "get": {
        "operationId": "GetEnterpriseGpResource",
        "summary": "Inspect Enterprise GP service or task",
        "description": "Inspect Enterprise GP service or task",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "resourceUrl",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "GP resource URL",
            "description": "Full National Grid ArcGIS Enterprise GPServer service or task URL."
          }
        ],
        "responses": {
          "200": {
            "description": "ArcGIS Enterprise JSON response.",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "x-ms-visibility": "important"
      }
    }
  },
...
}